verify x-signature
